+++ bpm-engine-impl/src/main/java/io/takari/bpm/state/StateHelper.java
@@ -8,11 +8,13 @@ import io.takari.bpm.actions.FireOnStartInterceptorsAction;
 import io.takari.bpm.api.ExecutionException;
 import io.takari.bpm.commands.CommandStack;
 import io.takari.bpm.commands.ProcessElementCommand;
+import io.takari.bpm.model.ProcessDefinition;
 import io.takari.bpm.model.StartEvent;
 
 import java.util.Arrays;
 import java.util.Map;
 import java.util.UUID;
+import java.util.function.Function;
 
 public final class StateHelper {
 
@@ -27,7 +29,7 @@ public final class StateHelper {
         state = state.setStack(stack.push(new ProcessElementCommand(pd.getId(), start.getId())));
 
         // set external variables
-        state = applyVariables(state, args);
+        state = applyVariables(state, pd.getAttributes(), args);
 
         // fire interceptors
         Action a = new FireOnStartInterceptorsAction(pd.getId());
@@ -36,14 +38,20 @@ public final class StateHelper {
         return state;
     }
 
-    public static ProcessInstance applyVariables(ProcessInstance state, Map<String, Object> args) {
-        if (args == null) {
-            return state;
+    public static ProcessInstance applyVariables(ProcessInstance state, Map<String, String> attr, Map<String, Object> args) {
+        Variables vars = state.getVariables();
+
+        if (attr != null) {
+            for (Map.Entry<String, String> e : attr.entrySet()) {
+                String k = ProcessDefinition.ATTRIBUTE_KEY_PREFIX + e.getKey();
+                vars = vars.setVariable(k, e.getValue());
+            }
         }
 
-        Variables vars = state.getVariables();
-        for (Map.Entry<String, Object> e : args.entrySet()) {
-            vars = vars.setVariable(e.getKey(), e.getValue());
+        if (args != null) {
+            for (Map.Entry<String, Object> e : args.entrySet()) {
+                vars = vars.setVariable(e.getKey(), e.getValue());
+            }
         }
 
         return state.setVariables(vars);